Bug fix: Bad check for last line of file
We were doing two things wrong.
First, we were loading the file for each iteration.
Second, we were checking against the value of the last line and assuming
the same value was a valid match.
This doesn't allow for the file to have duplicate values.

We now load the file as an iterator.
This lets us get the number of lines in the file.
We check against this line count to get the last line.